arch.py: use --visualize-jump to determine control flow
We use self defined function to determine control flow
in the program, this requires to have a full understanding
of the ISA and its state of the art features that might
be used for control flow.

From now on, we leverage the `--visualize-jump` option
in objdump to determine control flow in the program.
This is a more robust solution, as it does not require
a full understanding of the ISA, and it is more
maintainable in the long run.

Warning for users: Note that from now on,
we don't regard call instructions as control flow
instructions, as they are not considered as such
in the `--visualize-jump` output.

Signed-off-by: Yangyu Chen <cyy@cyyself.name>